% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/ChainBinomial-package.R
\docType{package}
\name{ChainBinomial-package}
\alias{ChainBinomial}
\alias{ChainBinomial-package}
\title{ChainBinomial: Chain Binomial Models for Infectious Disease Transmission}
\description{
Simulation and maximum likelihood estimation for chain binomial models of infectious disease transmission in household and community studies. Supports time-independent and time-dependent covariates, multiple incubation and infectious period distributions, and case-ascertained study designs.
}
\author{
\strong{Maintainer}: Yang Yang \email{yang.yang4@uga.edu}

}
\keyword{internal}
